feat(codebases): separate out components in different file and use tsx code
