+++ frontend/public/extend/devconsole/components/codebases/CodebaseHeader.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+import { ColHead, ListHeader } from '../../../../components/factory';
+
+const ComponentHeader = props => <ListHeader>
+  <ColHead {...props} className="col-lg-3 col-md-3 col-sm-4 col-xs-6" sortField="metadata.name">Name</ColHead>
+  <ColHead {...props} className="col-lg-2 col-md-2 col-sm-4 hidden-xs" sortField="metadata.namespace">Namespace</ColHead>
+  <ColHead {...props} className="col-lg-2 col-md-2 hidden-sm hidden-xs" sortField="spec.buildtype">BuildType</ColHead>
+  <ColHead {...props} className="col-lg-3 col-md-3 col-sm-4 col-xs-6" sortField="spec.codebase">Codebase</ColHead>
+  <ColHead {...props} className="col-lg-2 col-md-2 hidden-sm hidden-xs" sortField="spec.listenport">Port</ColHead>
+</ListHeader>;
+
+export default ComponentHeader;
