--- frontend/public/extend/devconsole/mockresources/component.yaml
@@ -1,119 +1,0 @@
-apiVersion: devopsconsole.openshift.io/v1alpha1
-kind: Component
-metadata:
-  name: christian
-  namespace: openshift
-spec:
-  # Add fields here
-  app: "openshiftio"
-  buildtype: "nodejs"
-  codebase: "https://github.com/sbose78/nodejs-rest-http-crud"
-  listenport: 8080
----
-apiVersion: devopsconsole.openshift.io/v1alpha1
-kind: Component
-metadata:
-  name: nimisha
-  namespace: openshift
-spec:
-  # Add fields here
-  app: "openshiftio"
-  buildtype: "react"
-  codebase: "https://github.com/sbose78/nodejs-rest-http-crud"
-  listenport: 9000
----
-apiVersion: devopsconsole.openshift.io/v1alpha1
-kind: Component
-metadata:
-  name: joshua
-  namespace: kube-dns
-spec:
-  # Add fields here
-  app: "kube-dns"
-  buildtype: "angular"
-  codebase: "https://github.com/sbose78/nodejs-rest-http-crud"
-  listenport: 4200
----
-apiVersion: devopsconsole.openshift.io/v1alpha1
-kind: Component
-metadata:
-  name: karthik
-  namespace: myproject
-spec:
-  # Add fields here
-  app: "myproject"
-  buildtype: "react"
-  codebase: "https://github.com/sbose78/nodejs-rest-http-crud"
-  listenport: 9000
----
-apiVersion: devopsconsole.openshift.io/v1alpha1
-kind: Component
-metadata:
-  name: debsmita
-  namespace: myproject
-spec:
-  # Add fields here
-  app: "myproject"
-  buildtype: "react"
-  codebase: "https://github.com/sbose78/nodejs-rest-http-crud"
-  listenport: 9000
----
-apiVersion: devopsconsole.openshift.io/v1alpha1
-kind: Component
-metadata:
-  name: rohit
-  namespace: kube-dns
-spec:
-  # Add fields here
-  app: "openshiftio"
-  buildtype: "nodejs"
-  codebase: "https://github.com/sbose78/nodejs-rest-http-crud"
-  listenport: 8080
----
-apiVersion: devopsconsole.openshift.io/v1alpha1
-kind: Component
-metadata:
-  name: sahil
-  namespace: kube-dns
-spec:
-  # Add fields here
-  app: "kube-dns"
-  buildtype: "react"
-  codebase: "https://github.com/sbose78/nodejs-rest-http-crud"
-  listenport: 9000
----
-apiVersion: devopsconsole.openshift.io/v1alpha1
-kind: Component
-metadata:
-  name: divyanshi
-  namespace: openshift
-spec:
-  # Add fields here
-  app: "openshiftio"
-  buildtype: "Angular"
-  codebase: "https://github.com/sbose78/nodejs-rest-http-crud"
-  listenport: 4200
-
----
-apiVersion: devopsconsole.openshift.io/v1alpha1
-kind: Component
-metadata:
-  name: vikram
-  namespace: openshift
-spec:
-  # Add fields here
-  app: "kube-dns"
-  buildtype: "nodejs"
-  codebase: "https://github.com/sbose78/nodejs-rest-http-crud"
-  listenport: 8080
----
-apiVersion: devopsconsole.openshift.io/v1alpha1
-kind: Component
-metadata:
-  name: dimo
-spec:
-  # Add fields here
-  app: "kube-dns"
-  buildtype: "nodejs"
-  codebase: "https://github.com/sbose78/nodejs-rest-http-crud"
-  listenport: 8080
