+++ frontend/public/extend/devconsole/mockresources/componentcrd.yaml
