+++ frontend/public/extend/devconsole/mockresources/components/debsmita.yaml
@@ -1,0 +1,10 @@
+apiVersion: devopsconsole.openshift.io/v1alpha1
+kind: Component
+metadata:
+  name: debsmita
+spec:
+  # Add fields here
+  app: "myproject"
+  buildtype: "react"
+  codebase: "https://github.com/sbose78/nodejs-rest-http-crud"
+  listenport: 9000
