+++ frontend/public/extend/devconsole/mockresources/components/sahil.yaml
@@ -1,0 +1,10 @@
+apiVersion: devopsconsole.openshift.io/v1alpha1
+kind: Component
+metadata:
+  name: sahil
+spec:
+  # Add fields here
+  app: "kube-dns"
+  buildtype: "react"
+  codebase: "https://github.com/sbose78/nodejs-rest-http-crud"
+  listenport: 9000
